requirements.txt
ccxt>=3.0.0
pandas>=1.3.0
numpy>=1.21.0
torch>=2.3.0
python-dotenv>=0.19.0
openai>=1.0.0
scikit-learn>=1.0.0
ta>=0.10.0
python-binance>=1.0.15
websockets>=10.0
aiohttp>=3.8.0
asyncio>=3.4.3
streamlit>=1.22.0
plotly>=5.13.0
matplotlib>=3.7.0
Pillow>=9.0.0
tensorboard>=2.14.0
stable-baselines3>=2.3.0
gym>=0.26.0
pyyaml>=6.0.2
pycocotools>=2.0.4
transformers>=4.36.0
